requirements.simple.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
PyMuPDF==1.23.8
pytesseract==0.3.10
gTTS==2.4.0
google-generativeai==0.3.2
requests==2.31.0
